feat: split template config in group and default
